refactor(tests): consolidate duplicated FakeCompletions test double
test_navigator_n2.py and test_navigator_n2_cookbooks.py (both added in #250)
each hand-rolled a byte-for-byte-equivalent FakeCompletions scripted chat
surface. Move it to conftest.py, the repo's existing home for shared test
fixtures/helpers, and import it from both call sites.
